which -> process.execPath

+++ testrunner.js
@@ -5,7 +5,6 @@ var Mocha = require('mocha');
 var path  = require('path');
 var fs    = require('fs');
 var spawn = require('child_process').spawn;
-var which = require('which');
 var argv  = require('yargs').argv;
 require('mocha-clean');
 require('intelli-espower-loader');
@@ -13,7 +12,7 @@ require('intelli-espower-loader');
 var mocha = new Mocha({
   ui: 'bdd',
   reporter: argv.R || 'spec',
-  timeout: 5000,
+  timeout: 20000,
   grep: argv.g
 });
 
@@ -22,18 +21,6 @@ fs.readdirSync(testDir).filter(function (file) {
   return /^[^_].*\.js$/.test(file);
 }).forEach(function (file) {
   mocha.addFile(path.join(testDir, file));
-});
-
-var server = spawn(which.sync('node'), [
-  path.join(__dirname, 'test/_server.js')
-], {
-  detached: true
-});
-server.stdout.on('data', function (data) {
-  process.stdout.write(data);
-});
-server.stderr.on('data', function (data) {
-  process.stderr.write(data);
 });
 
 // Ctrl-C
@@ -48,7 +35,21 @@ stdin.on('data', function (key) {
   }
 });
 
-mocha.run(function (failures) {
-  process.kill(server.pid);
-  process.exit(failures);
+var server = spawn(process.execPath, [
+  path.join(__dirname, 'test/_server.js')
+], {
+  detached: true
+});
+server.stdout.on('data', function (data) {
+  process.stdout.write(data);
+});
+server.stderr.on('data', function (data) {
+  if (data.toString() === '%%% server start %%%') {
+    // start mocha
+    return mocha.run(function (failures) {
+      process.kill(server.pid);
+      process.exit(failures);
+    });
+  }
+  process.stderr.write(data);
 });
